LPS-66183 Move template context creation logic to DDMFormTemplateContextFactoryImpl
