+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-form-renderer/src/main/java/com/liferay/dynamic/data/mapping/form/renderer/internal/DDMFormRendererImpl.java
@@ -14,61 +14,28 @@
 
 package com.liferay.dynamic.data.mapping.form.renderer.internal;
 
-import com.liferay.dynamic.data.mapping.form.evaluator.DDMFormEvaluationResult;
-import com.liferay.dynamic.data.mapping.form.evaluator.DDMFormEvaluator;
-import com.liferay.dynamic.data.mapping.form.field.type.DDMFormFieldType;
-import com.liferay.dynamic.data.mapping.form.field.type.DDMFormFieldTypeServicesTracker;
 import com.liferay.dynamic.data.mapping.form.renderer.DDMFormRenderer;
 import com.liferay.dynamic.data.mapping.form.renderer.DDMFormRenderingContext;
 import com.liferay.dynamic.data.mapping.form.renderer.DDMFormRenderingException;
-import com.liferay.dynamic.data.mapping.io.DDMFormFieldTypesJSONSerializer;
-import com.liferay.dynamic.data.mapping.io.DDMFormJSONSerializer;
-import com.liferay.dynamic.data.mapping.io.DDMFormLayoutJSONSerializer;
-import com.liferay.dynamic.data.mapping.io.DDMFormValuesJSONSerializer;
+import com.liferay.dynamic.data.mapping.form.renderer.DDMFormTemplateContextFactory;
 import com.liferay.dynamic.data.mapping.model.DDMForm;
-import com.liferay.dynamic.data.mapping.model.DDMFormField;
 import com.liferay.dynamic.data.mapping.model.DDMFormLayout;
-import com.liferay.dynamic.data.mapping.storage.DDMFormFieldValue;
-import com.liferay.dynamic.data.mapping.storage.DDMFormValues;
 import com.liferay.dynamic.data.mapping.util.DDM;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.io.unsync.UnsyncStringWriter;
-import com.liferay.portal.kernel.json.JSONArray;
-import com.liferay.portal.kernel.json.JSONFactory;
-import com.liferay.portal.kernel.json.JSONFactoryUtil;
-import com.liferay.portal.kernel.json.JSONSerializer;
-import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.template.Template;
 import com.liferay.portal.kernel.template.TemplateConstants;
 import com.liferay.portal.kernel.template.TemplateException;
 import com.liferay.portal.kernel.template.TemplateManagerUtil;
 import com.liferay.portal.kernel.template.TemplateResource;
 import com.liferay.portal.kernel.template.URLTemplateResource;
-import com.liferay.portal.kernel.util.AggregateResourceBundle;
-import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.MapUtil;
-import com.liferay.portal.kernel.util.PortalClassLoaderUtil;
-import com.liferay.portal.kernel.util.ResourceBundleUtil;
-import com.liferay.portal.kernel.util.StringBundler;
-import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.kernel.util.Validator;
 
 import java.io.Writer;
 
 import java.net.URL;
 
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Locale;
 import java.util.Map;
-import java.util.Objects;
-import java.util.ResourceBundle;
-
-import javax.servlet.Servlet;
-import javax.servlet.ServletConfig;
-import javax.servlet.ServletContext;
 
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
@@ -124,21 +91,6 @@ public class DDMFormRendererImpl implements DDMFormRenderer {
 		_templateResource = getTemplateResource(templatePath);
 	}
 
-	protected void collectResourceBundles(
-		Class<?> clazz, List<ResourceBundle> resourceBundles, Locale locale) {
-
-		for (Class<?> interfaceClass : clazz.getInterfaces()) {
-			collectResourceBundles(interfaceClass, resourceBundles, locale);
-		}
-
-		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
-			"content.Language", locale, clazz.getClassLoader());
-
-		if (resourceBundle != null) {
-			resourceBundles.add(resourceBundle);
-		}
-	}
-
 	protected String doRender(
 			DDMForm ddmForm, DDMFormLayout ddmFormLayout,
 			DDMFormRenderingContext ddmFormRenderingContext)
@@ -150,136 +102,13 @@ public class DDMFormRendererImpl implements DDMFormRenderer {
 		populateCommonContext(
 			template, ddmForm, ddmFormLayout, ddmFormRenderingContext);
 
-		String html = render(template, getTemplateNamespace(ddmFormLayout));
+		String templateNamespace = (String)template.get("templateNamespace");
+
+		String html = render(template, templateNamespace);
 
 		String javaScript = render(template, "ddm.form_renderer_js");
 
 		return html.concat(javaScript);
-	}
-
-	protected String getDDMDataProviderServletURL() {
-		String servletContextPath = getServletContextPath(
-			_ddmDataProviderServlet);
-
-		return servletContextPath.concat(
-			"/dynamic-data-mapping-data-provider/");
-	}
-
-	protected String getDDMFormEvaluatorServletURL() {
-		String servletContextPath = getServletContextPath(
-			_ddmFormEvaluatorServlet);
-
-		return servletContextPath.concat(
-			"/dynamic-data-mapping-form-evaluator/");
-	}
-
-	protected Map<String, String> getLanguageStringsMap(
-		ResourceBundle resourceBundle) {
-
-		Map<String, String> stringsMap = new HashMap<>();
-
-		stringsMap.put("next", LanguageUtil.get(resourceBundle, "next"));
-		stringsMap.put(
-			"previous", LanguageUtil.get(resourceBundle, "previous"));
-
-		return stringsMap;
-	}
-
-	protected List<Object> getPages(
-			DDMForm ddmForm, DDMFormLayout ddmFormLayout,
-			DDMFormRenderingContext ddmFormRenderingContext)
-		throws DDMFormRenderingException {
-
-		Map<String, String> renderedDDMFormFieldsMap =
-			getRenderedDDMFormFieldsMap(ddmForm, ddmFormRenderingContext);
-
-		DDMFormLayoutTransformer ddmFormLayoutTransformer =
-			new DDMFormLayoutTransformer(
-				ddmForm, ddmFormLayout, renderedDDMFormFieldsMap,
-				ddmFormRenderingContext.isShowRequiredFieldsWarning(),
-				ddmFormRenderingContext.getLocale());
-
-		return ddmFormLayoutTransformer.getPages();
-	}
-
-	protected JSONArray getReadOnlyFieldsJSONArray(DDMForm ddmForm) {
-		JSONArray jsonArray = JSONFactoryUtil.createJSONArray();
-
-		List<DDMFormField> ddmFormFields = ddmForm.getDDMFormFields();
-
-		for (DDMFormField ddmFormField : ddmFormFields) {
-			if (ddmFormField.isReadOnly()) {
-				jsonArray.put(ddmFormField.getName());
-			}
-		}
-
-		return jsonArray;
-	}
-
-	protected Map<String, String> getRenderedDDMFormFieldsMap(
-			DDMForm ddmForm, DDMFormRenderingContext ddmFormRenderingContext)
-		throws DDMFormRenderingException {
-
-		DDMFormRendererHelper ddmFormRendererHelper = new DDMFormRendererHelper(
-			ddmForm, ddmFormRenderingContext);
-
-		ddmFormRendererHelper.setDDMFormFieldTypeServicesTracker(
-			_ddmFormFieldTypeServicesTracker);
-		ddmFormRendererHelper.setDDMFormEvaluator(_ddmFormEvaluator);
-
-		return ddmFormRendererHelper.getRenderedDDMFormFieldsMap();
-	}
-
-	protected String getRequiredFieldsWarningMessageHTML(
-		ResourceBundle resourceBundle) {
-
-		StringBundler sb = new StringBundler(3);
-
-		sb.append("<label class=\"required-warning\">");
-		sb.append(
-			LanguageUtil.format(
-				resourceBundle, "all-fields-marked-with-x-are-required",
-				"<i class=\"icon-asterisk text-warning\"></i>", false));
-		sb.append("</label>");
-
-		return sb.toString();
-	}
-
-	protected ResourceBundle getResourceBundle(Locale locale) {
-		List<ResourceBundle> resourceBundles = new ArrayList<>();
-
-		ResourceBundle portalResourceBundle = ResourceBundleUtil.getBundle(
-			"content.Language", locale, PortalClassLoaderUtil.getClassLoader());
-
-		resourceBundles.add(portalResourceBundle);
-
-		collectResourceBundles(getClass(), resourceBundles, locale);
-
-		ResourceBundle[] resourceBundlesArray = resourceBundles.toArray(
-			new ResourceBundle[resourceBundles.size()]);
-
-		return new AggregateResourceBundle(resourceBundlesArray);
-	}
-
-	protected String getServletContextPath(Servlet servlet) {
-		ServletConfig servletConfig = servlet.getServletConfig();
-
-		ServletContext servletContext = servletConfig.getServletContext();
-
-		return servletContext.getContextPath();
-	}
-
-	protected String getTemplateNamespace(DDMFormLayout ddmFormLayout) {
-		String paginationMode = ddmFormLayout.getPaginationMode();
-
-		if (Objects.equals(paginationMode, DDMFormLayout.SINGLE_PAGE_MODE)) {
-			return "ddm.simple_form";
-		}
-		else if (Objects.equals(paginationMode, DDMFormLayout.TABBED_MODE)) {
-			return "ddm.tabbed_form";
-		}
-
-		return "ddm.paginated_form";
 	}
 
 	protected TemplateResource getTemplateResource(String templatePath) {
@@ -297,114 +126,11 @@ public class DDMFormRendererImpl implements DDMFormRenderer {
 			DDMFormRenderingContext ddmFormRenderingContext)
 		throws PortalException {
 
-		String containerId = ddmFormRenderingContext.getContainerId();
+		Map<String, Object> templateContext =
+			_ddmFormTemplateContextFactory.create(
+				ddmForm, ddmFormLayout, ddmFormRenderingContext);
 
-		if (Validator.isNull(containerId)) {
-			containerId = StringUtil.randomId();
-		}
-
-		template.put("containerId", containerId);
-		template.put("dataProviderURL", getDDMDataProviderServletURL());
-		template.put("definition", _ddmFormJSONSerializer.serialize(ddmForm));
-
-		DDMFormValues ddmFormValues =
-			ddmFormRenderingContext.getDDMFormValues();
-
-		if (ddmFormValues != null) {
-			removeStaleDDMFormFieldValues(
-				ddmForm.getDDMFormFieldsMap(true),
-				ddmFormValues.getDDMFormFieldValues());
-		}
-
-		Locale locale = ddmFormRenderingContext.getLocale();
-
-		DDMFormEvaluationResult ddmFormEvaluationResult =
-			_ddmFormEvaluator.evaluate(
-				ddmForm, ddmFormRenderingContext.getDDMFormValues(), locale);
-
-		JSONSerializer jsonSerializer = _jsonFactory.createJSONSerializer();
-
-		template.put(
-			"evaluation",
-			jsonSerializer.serializeDeep(ddmFormEvaluationResult));
-		template.put("evaluatorURL", getDDMFormEvaluatorServletURL());
-
-		List<DDMFormFieldType> ddmFormFieldTypes =
-			_ddmFormFieldTypeServicesTracker.getDDMFormFieldTypes();
-
-		template.put(
-			"fieldTypes",
-			_ddmFormFieldTypesJSONSerializer.serialize(ddmFormFieldTypes));
-		template.put(
-			"layout", _ddmFormLayoutJSONSerializer.serialize(ddmFormLayout));
-
-		List<Object> pages = getPages(
-			ddmForm, ddmFormLayout, ddmFormRenderingContext);
-
-		template.put("pages", pages);
-
-		template.put(
-			"portletNamespace", ddmFormRenderingContext.getPortletNamespace());
-		template.put("readOnly", ddmFormRenderingContext.isReadOnly());
-
-		JSONArray readOnlyFieldsJSONArray = getReadOnlyFieldsJSONArray(ddmForm);
-
-		template.put("readOnlyFields", readOnlyFieldsJSONArray.toString());
-
-		ResourceBundle resourceBundle = getResourceBundle(locale);
-
-		template.put(
-			"requiredFieldsWarningMessageHTML",
-			getRequiredFieldsWarningMessageHTML(resourceBundle));
-		template.put(
-			"showRequiredFieldsWarning",
-			ddmFormRenderingContext.isShowRequiredFieldsWarning());
-
-		boolean showSubmitButton = ddmFormRenderingContext.isShowSubmitButton();
-
-		if (ddmFormRenderingContext.isReadOnly()) {
-			showSubmitButton = false;
-		}
-
-		template.put("showSubmitButton", showSubmitButton);
-
-		template.put("strings", getLanguageStringsMap(resourceBundle));
-
-		String submitLabel = GetterUtil.getString(
-			ddmFormRenderingContext.getSubmitLabel(),
-			LanguageUtil.get(locale, "submit"));
-
-		template.put("submitLabel", submitLabel);
-
-		template.put("templateNamespace", getTemplateNamespace(ddmFormLayout));
-
-		if (ddmFormValues != null) {
-			template.put(
-				"values",
-				_ddmFormValuesJSONSerializer.serialize(ddmFormValues));
-		}
-		else {
-			template.put("values", JSONFactoryUtil.getNullJSON());
-		}
-	}
-
-	protected void removeStaleDDMFormFieldValues(
-		Map<String, DDMFormField> ddmFormFieldsMap,
-		List<DDMFormFieldValue> ddmFormFieldValues) {
-
-		Iterator<DDMFormFieldValue> iterator = ddmFormFieldValues.iterator();
-
-		while (iterator.hasNext()) {
-			DDMFormFieldValue ddmFormFieldValue = iterator.next();
-
-			if (!ddmFormFieldsMap.containsKey(ddmFormFieldValue.getName())) {
-				iterator.remove();
-			}
-
-			removeStaleDDMFormFieldValues(
-				ddmFormFieldsMap,
-				ddmFormFieldValue.getNestedDDMFormFieldValues());
-		}
+		template.putAll(templateContext);
 	}
 
 	protected String render(Template template, String namespace)
@@ -425,77 +151,12 @@ public class DDMFormRendererImpl implements DDMFormRenderer {
 		_ddm = ddm;
 	}
 
-	@Reference(
-		target = "(osgi.http.whiteboard.servlet.name=com.liferay.dynamic.data.mapping.data.provider.internal.servlet.DDMDataProviderServlet)",
-		unbind = "-"
-	)
-	protected void setDDMDataProviderServlet(Servlet ddmDataProviderServlet) {
-		_ddmDataProviderServlet = ddmDataProviderServlet;
-	}
-
-	@Reference(unbind = "-")
-	protected void setDDMFormEvaluator(DDMFormEvaluator ddmFormEvaluator) {
-		_ddmFormEvaluator = ddmFormEvaluator;
-	}
-
-	@Reference(
-		target = "(osgi.http.whiteboard.servlet.name=com.liferay.dynamic.data.mapping.form.evaluator.internal.servlet.DDMFormEvaluatorServlet)",
-		unbind = "-"
-	)
-	protected void setDDMFormEvaluatorServlet(Servlet ddmFormEvaluatorServlet) {
-		_ddmFormEvaluatorServlet = ddmFormEvaluatorServlet;
-	}
-
-	@Reference(unbind = "-")
-	protected void setDDMFormFieldTypeServicesTracker(
-		DDMFormFieldTypeServicesTracker ddmFormFieldTypeServicesTracker) {
-
-		_ddmFormFieldTypeServicesTracker = ddmFormFieldTypeServicesTracker;
-	}
-
-	@Reference(unbind = "-")
-	protected void setDDMFormFieldTypesJSONSerializer(
-		DDMFormFieldTypesJSONSerializer ddmFormFieldTypesJSONSerializer) {
-
-		_ddmFormFieldTypesJSONSerializer = ddmFormFieldTypesJSONSerializer;
-	}
-
-	@Reference(unbind = "-")
-	protected void setDDMFormJSONSerializer(
-		DDMFormJSONSerializer ddmFormJSONSerializer) {
-
-		_ddmFormJSONSerializer = ddmFormJSONSerializer;
-	}
-
-	@Reference(unbind = "-")
-	protected void setDDMFormLayoutJSONSerializer(
-		DDMFormLayoutJSONSerializer ddmFormLayoutJSONSerializer) {
-
-		_ddmFormLayoutJSONSerializer = ddmFormLayoutJSONSerializer;
-	}
-
-	@Reference(unbind = "-")
-	protected void setDDMFormValuesJSONSerializer(
-		DDMFormValuesJSONSerializer ddmFormValuesJSONSerializer) {
-
-		_ddmFormValuesJSONSerializer = ddmFormValuesJSONSerializer;
-	}
-
-	@Reference(unbind = "-")
-	protected void setJSONFactory(JSONFactory jsonFactory) {
-		_jsonFactory = jsonFactory;
-	}
-
+	@Reference
 	private DDM _ddm;
-	private Servlet _ddmDataProviderServlet;
-	private DDMFormEvaluator _ddmFormEvaluator;
-	private Servlet _ddmFormEvaluatorServlet;
-	private DDMFormFieldTypeServicesTracker _ddmFormFieldTypeServicesTracker;
-	private DDMFormFieldTypesJSONSerializer _ddmFormFieldTypesJSONSerializer;
-	private DDMFormJSONSerializer _ddmFormJSONSerializer;
-	private DDMFormLayoutJSONSerializer _ddmFormLayoutJSONSerializer;
-	private DDMFormValuesJSONSerializer _ddmFormValuesJSONSerializer;
-	private JSONFactory _jsonFactory;
+
+	@Reference
+	private DDMFormTemplateContextFactory _ddmFormTemplateContextFactory;
+
 	private TemplateResource _templateResource;
 
 }
