+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-form-renderer/src/main/java/com/liferay/dynamic/data/mapping/form/renderer/internal/DDMFormTemplateContextFactoryImpl.java
@@ -14,15 +14,52 @@
 
 package com.liferay.dynamic.data.mapping.form.renderer.internal;
 
+import com.liferay.dynamic.data.mapping.form.evaluator.DDMFormEvaluationResult;
+import com.liferay.dynamic.data.mapping.form.evaluator.DDMFormEvaluator;
+import com.liferay.dynamic.data.mapping.form.field.type.DDMFormFieldType;
+import com.liferay.dynamic.data.mapping.form.field.type.DDMFormFieldTypeServicesTracker;
 import com.liferay.dynamic.data.mapping.form.renderer.DDMFormRenderingContext;
+import com.liferay.dynamic.data.mapping.form.renderer.DDMFormRenderingException;
 import com.liferay.dynamic.data.mapping.form.renderer.DDMFormTemplateContextFactory;
+import com.liferay.dynamic.data.mapping.io.DDMFormFieldTypesJSONSerializer;
+import com.liferay.dynamic.data.mapping.io.DDMFormJSONSerializer;
+import com.liferay.dynamic.data.mapping.io.DDMFormLayoutJSONSerializer;
+import com.liferay.dynamic.data.mapping.io.DDMFormValuesJSONSerializer;
 import com.liferay.dynamic.data.mapping.model.DDMForm;
+import com.liferay.dynamic.data.mapping.model.DDMFormField;
 import com.liferay.dynamic.data.mapping.model.DDMFormLayout;
+import com.liferay.dynamic.data.mapping.storage.DDMFormFieldValue;
+import com.liferay.dynamic.data.mapping.storage.DDMFormValues;
+import com.liferay.dynamic.data.mapping.util.DDM;
 import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.json.JSONArray;
+import com.liferay.portal.kernel.json.JSONFactory;
+import com.liferay.portal.kernel.json.JSONFactoryUtil;
+import com.liferay.portal.kernel.json.JSONSerializer;
+import com.liferay.portal.kernel.language.LanguageUtil;
+import com.liferay.portal.kernel.util.AggregateResourceBundle;
+import com.liferay.portal.kernel.util.GetterUtil;
+import com.liferay.portal.kernel.util.PortalClassLoaderUtil;
+import com.liferay.portal.kernel.util.ResourceBundleUtil;
+import com.liferay.portal.kernel.util.StringBundler;
+import com.liferay.portal.kernel.util.StringUtil;
+import com.liferay.portal.kernel.util.Validator;
 
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Locale;
 import java.util.Map;
+import java.util.Objects;
+import java.util.ResourceBundle;
+
+import javax.servlet.Servlet;
+import javax.servlet.ServletConfig;
+import javax.servlet.ServletContext;
 
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Marcellus Tavares
@@ -36,14 +73,314 @@ public class DDMFormTemplateContextFactoryImpl
 			DDMFormRenderingContext ddmFormRenderingContext)
 		throws PortalException {
 
-		return null;
+		return doCreate(ddmForm, ddmFormLayout, ddmFormRenderingContext);
 	}
 
 	public Map<String, Object> create(
 			DDMForm ddmForm, DDMFormRenderingContext ddmFormRenderingContext)
 		throws PortalException {
 
-		return null;
+		return doCreate(
+			ddmForm, _ddm.getDefaultDDMFormLayout(ddmForm),
+			ddmFormRenderingContext);
 	}
+
+	protected void collectResourceBundles(
+		Class<?> clazz, List<ResourceBundle> resourceBundles, Locale locale) {
+
+		for (Class<?> interfaceClass : clazz.getInterfaces()) {
+			collectResourceBundles(interfaceClass, resourceBundles, locale);
+		}
+
+		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
+			"content.Language", locale, clazz.getClassLoader());
+
+		if (resourceBundle != null) {
+			resourceBundles.add(resourceBundle);
+		}
+	}
+
+	protected Map<String, Object> doCreate(
+			DDMForm ddmForm, DDMFormLayout ddmFormLayout,
+			DDMFormRenderingContext ddmFormRenderingContext)
+		throws PortalException {
+
+		Map<String, Object> templateContext = new HashMap<>();
+
+		String containerId = ddmFormRenderingContext.getContainerId();
+
+		if (Validator.isNull(containerId)) {
+			containerId = StringUtil.randomId();
+		}
+
+		templateContext.put("containerId", containerId);
+		templateContext.put("dataProviderURL", getDDMDataProviderServletURL());
+		templateContext.put(
+			"definition", _ddmFormJSONSerializer.serialize(ddmForm));
+
+		DDMFormValues ddmFormValues =
+			ddmFormRenderingContext.getDDMFormValues();
+
+		if (ddmFormValues != null) {
+			removeStaleDDMFormFieldValues(
+				ddmForm.getDDMFormFieldsMap(true),
+				ddmFormValues.getDDMFormFieldValues());
+		}
+
+		Locale locale = ddmFormRenderingContext.getLocale();
+
+		DDMFormEvaluationResult ddmFormEvaluationResult =
+			_ddmFormEvaluator.evaluate(
+				ddmForm, ddmFormRenderingContext.getDDMFormValues(), locale);
+
+		JSONSerializer jsonSerializer = _jsonFactory.createJSONSerializer();
+
+		templateContext.put(
+			"evaluation",
+			jsonSerializer.serializeDeep(ddmFormEvaluationResult));
+		templateContext.put("evaluatorURL", getDDMFormEvaluatorServletURL());
+
+		List<DDMFormFieldType> ddmFormFieldTypes =
+			_ddmFormFieldTypeServicesTracker.getDDMFormFieldTypes();
+
+		templateContext.put(
+			"fieldTypes",
+			_ddmFormFieldTypesJSONSerializer.serialize(ddmFormFieldTypes));
+		templateContext.put(
+			"layout", _ddmFormLayoutJSONSerializer.serialize(ddmFormLayout));
+
+		List<Object> pages = getPages(
+			ddmForm, ddmFormLayout, ddmFormRenderingContext);
+
+		templateContext.put("pages", pages);
+
+		templateContext.put(
+			"portletNamespace", ddmFormRenderingContext.getPortletNamespace());
+		templateContext.put("readOnly", ddmFormRenderingContext.isReadOnly());
+
+		JSONArray readOnlyFieldsJSONArray = getReadOnlyFieldsJSONArray(ddmForm);
+
+		templateContext.put(
+			"readOnlyFields", readOnlyFieldsJSONArray.toString());
+
+		ResourceBundle resourceBundle = getResourceBundle(locale);
+
+		templateContext.put(
+			"requiredFieldsWarningMessageHTML",
+			getRequiredFieldsWarningMessageHTML(resourceBundle));
+		templateContext.put(
+			"showRequiredFieldsWarning",
+			ddmFormRenderingContext.isShowRequiredFieldsWarning());
+
+		boolean showSubmitButton = ddmFormRenderingContext.isShowSubmitButton();
+
+		if (ddmFormRenderingContext.isReadOnly()) {
+			showSubmitButton = false;
+		}
+
+		templateContext.put("showSubmitButton", showSubmitButton);
+
+		templateContext.put("strings", getLanguageStringsMap(resourceBundle));
+
+		String submitLabel = GetterUtil.getString(
+			ddmFormRenderingContext.getSubmitLabel(),
+			LanguageUtil.get(locale, "submit"));
+
+		templateContext.put("submitLabel", submitLabel);
+
+		templateContext.put(
+			"templateNamespace", getTemplateNamespace(ddmFormLayout));
+
+		if (ddmFormValues != null) {
+			templateContext.put(
+				"values",
+				_ddmFormValuesJSONSerializer.serialize(ddmFormValues));
+		}
+		else {
+			templateContext.put("values", JSONFactoryUtil.getNullJSON());
+		}
+
+		return templateContext;
+	}
+
+	protected String getDDMDataProviderServletURL() {
+		String servletContextPath = getServletContextPath(
+			_ddmDataProviderServlet);
+
+		return servletContextPath.concat(
+			"/dynamic-data-mapping-data-provider/");
+	}
+
+	protected String getDDMFormEvaluatorServletURL() {
+		String servletContextPath = getServletContextPath(
+			_ddmFormEvaluatorServlet);
+
+		return servletContextPath.concat(
+			"/dynamic-data-mapping-form-evaluator/");
+	}
+
+	protected Map<String, String> getLanguageStringsMap(
+		ResourceBundle resourceBundle) {
+
+		Map<String, String> stringsMap = new HashMap<>();
+
+		stringsMap.put("next", LanguageUtil.get(resourceBundle, "next"));
+		stringsMap.put(
+			"previous", LanguageUtil.get(resourceBundle, "previous"));
+
+		return stringsMap;
+	}
+
+	protected List<Object> getPages(
+			DDMForm ddmForm, DDMFormLayout ddmFormLayout,
+			DDMFormRenderingContext ddmFormRenderingContext)
+		throws DDMFormRenderingException {
+
+		Map<String, String> renderedDDMFormFieldsMap =
+			getRenderedDDMFormFieldsMap(ddmForm, ddmFormRenderingContext);
+
+		DDMFormLayoutTransformer ddmFormLayoutTransformer =
+			new DDMFormLayoutTransformer(
+				ddmForm, ddmFormLayout, renderedDDMFormFieldsMap,
+				ddmFormRenderingContext.isShowRequiredFieldsWarning(),
+				ddmFormRenderingContext.getLocale());
+
+		return ddmFormLayoutTransformer.getPages();
+	}
+
+	protected JSONArray getReadOnlyFieldsJSONArray(DDMForm ddmForm) {
+		JSONArray jsonArray = JSONFactoryUtil.createJSONArray();
+
+		List<DDMFormField> ddmFormFields = ddmForm.getDDMFormFields();
+
+		for (DDMFormField ddmFormField : ddmFormFields) {
+			if (ddmFormField.isReadOnly()) {
+				jsonArray.put(ddmFormField.getName());
+			}
+		}
+
+		return jsonArray;
+	}
+
+	protected Map<String, String> getRenderedDDMFormFieldsMap(
+			DDMForm ddmForm, DDMFormRenderingContext ddmFormRenderingContext)
+		throws DDMFormRenderingException {
+
+		DDMFormRendererHelper ddmFormRendererHelper = new DDMFormRendererHelper(
+			ddmForm, ddmFormRenderingContext);
+
+		ddmFormRendererHelper.setDDMFormFieldTypeServicesTracker(
+			_ddmFormFieldTypeServicesTracker);
+		ddmFormRendererHelper.setDDMFormEvaluator(_ddmFormEvaluator);
+
+		return ddmFormRendererHelper.getRenderedDDMFormFieldsMap();
+	}
+
+	protected String getRequiredFieldsWarningMessageHTML(
+		ResourceBundle resourceBundle) {
+
+		StringBundler sb = new StringBundler(3);
+
+		sb.append("<label class=\"required-warning\">");
+		sb.append(
+			LanguageUtil.format(
+				resourceBundle, "all-fields-marked-with-x-are-required",
+				"<i class=\"icon-asterisk text-warning\"></i>", false));
+		sb.append("</label>");
+
+		return sb.toString();
+	}
+
+	protected ResourceBundle getResourceBundle(Locale locale) {
+		List<ResourceBundle> resourceBundles = new ArrayList<>();
+
+		ResourceBundle portalResourceBundle = ResourceBundleUtil.getBundle(
+			"content.Language", locale, PortalClassLoaderUtil.getClassLoader());
+
+		resourceBundles.add(portalResourceBundle);
+
+		collectResourceBundles(getClass(), resourceBundles, locale);
+
+		ResourceBundle[] resourceBundlesArray = resourceBundles.toArray(
+			new ResourceBundle[resourceBundles.size()]);
+
+		return new AggregateResourceBundle(resourceBundlesArray);
+	}
+
+	protected String getServletContextPath(Servlet servlet) {
+		ServletConfig servletConfig = servlet.getServletConfig();
+
+		ServletContext servletContext = servletConfig.getServletContext();
+
+		return servletContext.getContextPath();
+	}
+
+	protected String getTemplateNamespace(DDMFormLayout ddmFormLayout) {
+		String paginationMode = ddmFormLayout.getPaginationMode();
+
+		if (Objects.equals(paginationMode, DDMFormLayout.SINGLE_PAGE_MODE)) {
+			return "ddm.simple_form";
+		}
+		else if (Objects.equals(paginationMode, DDMFormLayout.TABBED_MODE)) {
+			return "ddm.tabbed_form";
+		}
+
+		return "ddm.paginated_form";
+	}
+
+	protected void removeStaleDDMFormFieldValues(
+		Map<String, DDMFormField> ddmFormFieldsMap,
+		List<DDMFormFieldValue> ddmFormFieldValues) {
+
+		Iterator<DDMFormFieldValue> iterator = ddmFormFieldValues.iterator();
+
+		while (iterator.hasNext()) {
+			DDMFormFieldValue ddmFormFieldValue = iterator.next();
+
+			if (!ddmFormFieldsMap.containsKey(ddmFormFieldValue.getName())) {
+				iterator.remove();
+			}
+
+			removeStaleDDMFormFieldValues(
+				ddmFormFieldsMap,
+				ddmFormFieldValue.getNestedDDMFormFieldValues());
+		}
+	}
+
+	@Reference
+	private DDM _ddm;
+
+	@Reference(
+		target = "(osgi.http.whiteboard.servlet.name=com.liferay.dynamic.data.mapping.data.provider.internal.servlet.DDMDataProviderServlet)",
+		unbind = "-"
+	)
+	private Servlet _ddmDataProviderServlet;
+
+	@Reference
+	private DDMFormEvaluator _ddmFormEvaluator;
+
+	@Reference(
+		target = "(osgi.http.whiteboard.servlet.name=com.liferay.dynamic.data.mapping.form.evaluator.internal.servlet.DDMFormEvaluatorServlet)",
+		unbind = "-"
+	)
+	private Servlet _ddmFormEvaluatorServlet;
+
+	@Reference
+	private DDMFormFieldTypeServicesTracker _ddmFormFieldTypeServicesTracker;
+
+	@Reference
+	private DDMFormFieldTypesJSONSerializer _ddmFormFieldTypesJSONSerializer;
+
+	@Reference
+	private DDMFormJSONSerializer _ddmFormJSONSerializer;
+
+	@Reference
+	private DDMFormLayoutJSONSerializer _ddmFormLayoutJSONSerializer;
+
+	@Reference
+	private DDMFormValuesJSONSerializer _ddmFormValuesJSONSerializer;
+
+	@Reference
+	private JSONFactory _jsonFactory;
 
 }
